Make the API for dao more ffluent and transparent
First change done is separated the Read and write interfaces. Second
change done is name for the read APIs made more consistent. Introduce
common daos' for read and read-write.
This changes were inspired by the fact that caching will be introduced
in the API.

Add .gitignore file to ignore contents of target folder

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>