Add basic files for testing smart dao
It will use Apache Derby DB in embedded mode, along with Spring and
Hibernate. For configuring the db to use one may use the db-test-config
template properties file and to get it to work it should work if only
copy and paste is done.

Three test domains - Book, Author and Publisher along with their
hibernate mapping file with associations.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>