Add the new Query Parameter API and SPI
The new parameter API is designed on the basis of type and number of
parameters required for constituting a valid parameter. Basic emphasis
was on interface based design to hide the implementation details.

It is not yet integrated into QueryParameterFactory and AbstractDao,
that will be the next 2 stages followed by integration with test.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>