Fix query parameters' concrete implementation
Previously the query parameters where using a base abstract class to
perform the operations of different types of query parameters. This was
causing erroneous "instanceof" output which was not a good API design.
Thus this change fixes that problem and ensures that "instanceof" works
as expected.

Appropriate changes are made to the hibernate implementation to conform
with this change.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>