Refactor the domain package to a separate module
Previously domain was a part of the abstract-dao module and thus to
ensure that a project follows the smart-dao domain model it was needed
to import the whole module instead of the domain part only, thus this
refactor was introduced.

From this refactor another benefit that can be reaped is domain export,
validation checking and binding frameworks basic work can be achieved.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>