Implement DB based version search
Firstly, changed the search operation signatures, search can now be
performed on commits and revisions only.

Secondly, added the search properties enum so that its easy to know the
searchable properties.

Thirdly, domain to API operations in meta factory is made static and its
a FIX.

Fourthly, made appropriate changes to meta service API and
implementation. The implementation is simple as it can blindly transfer
the query parameters to the reader daos. Its main responsibility is
transformation and java.util.Set creation. Users using search should
take a look at the tests that will be added in next commit to ensure they
are using the params properly.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>