Refactor Version API and Implementation
This refactor is mainly done to bring about a more mature API and
implementation.

Previously it resources did not supported whether they are deleted or not
and as a result the API was not able to handle deleted resource, but
this change ensures that resources are containers of deleted flag. Thus
now there is a unified write operation.

Previously content type of resource was saved, but now the API accepts it
as an optional attribute, it would be useful when we will present
resources through RESTful Web Services.

Previously there wasn't any relation between SPI and API but in reality
all SPI should extend the API as well and that is ensured through
extension.

Previously read and write dao for version was a single one, but now it
is changed to be different.

Previously default config did not allow read and write location to be
distinct, but now the config will allow it.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>