Add annotations API for exporting domains in RESTful manner
It is a simple API for marking information useful for exporting objects.
Its main purpose lies in exporting in REST manner, with main
concentration on URI based object export.

ResourceDomain is an annotation intended for domain/resource classes to
be exported via REST. If the domain is not the JAX-RS resource then it
should mention the path to the resource so that when its associated with
other resource the path can be used. Also client can specify how to
export the domain by defaut and its URI by default. Access mechanism to
attributes can also be specified and current supported access mechanisms
are field and property accessor and the latter is enabled by default.

Id is intended to mark resource identifying attribute in case of export
as URI and a prefix path can also be specified in this case.

Export is to specify whether to export or not if yes then how, as
an object or URI; by default its the latter.

Name signifies the name representing an attirbute of the resource, if
not mentioned it will use the name of the property name.

Eager signifies that the associated resource is to be loaded
aggresively, if not mentioned it will be loaded on first access of an
non-id property.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>