Update the XSL to produce more rich log
With this update the commits point to the web pages with more details to
the changes, namingly to github and repo.or.cz

Relocate google host pages under gh-pages directory.

Added a changes aggregating HTML, that is intended to contain/embed changes
for all releases.

Also adds the CSS file to the head of the HTML and to all outputted values.
Also add the corresponding CSS.

Plus the XSL now handles text output properly, i.e. wraps in <xsl:text>

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>