Make distribution more complete
Previously the modules did not have any READMEs', this change introduces
README for every module and also includes them in the distros.

The READMEs are based on the fields in POM and are filtered at module level
to ensure that proper information is populated.

The with dependency assembly descriptor is minimized by excluding the
inclusion manually and rather configuring regex based exclusion filter and
thus ensuring all dependency is included.

Docs is now registered as a module directly rather than from within a
profile since all costly operations are performed in profiles which need to
be enabled manually.

The dists & dependency are not structured module wise becuase of inter
module dependency and which would cause duplication of JARs thus the
users are advised to either use maven or consult the module dependency page
in the docs.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>