All distribution is now built under the same profile
Since building distribution is auxiliary to building the site its moved to
the unique profile which is intended only for it. Also the dist artifacts
are marked as non-attchs', i.e. they are not artifacts to be tracked by
maven.

Also changed the packaging of the doc module to POM only since its
artifacts are non useful.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>