Add the preliminary API for RESTful import-export of objects
It contains the implementation of a JAX-RS provider, both producer and
consumer. It also adds the basic tests of the provider.

For the JAX-RS provider it adds a import export (EXIM) API that will bridge
between the provider and implementations of the different media type
handlers.

The EXIM API lacks pre/post processors which will be added only after the
basic media type supports are implemented.

Jersey is added to the test classpath so that JAX-RS delegation works as it
should. Also import is fixed for a cache test case.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>