Refactor smart-rs API into 2 modules
Smart-RS will now be concerned with JAX-RS related functionalities only.

The newly introduced Smart-ExIm will be concerned with export and import of
resources with respect to their URIs. It is independent of JAX-RS or any
other RESTful framework for that matter. Smart ExIm will be broken down
into API, which will contain the common code base and ExIm provider modules
concerned with specific media type export and import.

Also updated Jersey version to the stable one.

Also fixed the site page title of a HTML page.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>