Load classes for default implementation of scanners
So that default instances of scanners are available at all time and the
default instances are dependent on static initializers, the registrar scans
for the default scanners in the "impl" package relative to it and loads the
classes so that the static initializers are loaded.

Also added a test to verify it.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>