Make small fixes
The first fix is to make first character of the property name lower case
instead of upper case.

The second fix is to check whether resource domain is null or not, if null
then ignore it.

The third fix is check for configuration in already scanned configurations.

The fourth fix is use factory for generating searchable annotation name.

The fifth change is using factory for getting the class of probably
resource rather than doing it manually.

Also add the license header to the test file.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>