Fix registrar item initialization
Previously when the item was initializing it was checking for getInstance
method to be static using isSythetic method which is wrong, now it is fixed
by using Modifier.isStatic instead.

Test case was missing for these case thus added one for that as well, to
ensure that this does not go unchecked in future.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>