Specify the SPI for HBase DAO
Conversion of instance to Put/Delete and Result to instance is delegated
to ObjectRowConverter. Which could either be an Object-BigTable Mapping
implementation or simple implementation to handle a specific type.

SchemaInfoProvider is there to provide schema related information so that
we can use them to convert query params to filter, also class to its
respective tables etc.

Filter config is added to provide a column specific filter ops configs.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>