Add simple query param API based read implementation in DAO
Currently only a very small subset of the query API is implemented. Only
operations on a property, i.e. cell, is supported comprehensively enough to
be compared to the SQL implementation. Conjunction and disjunction is also
supported quite comprehensively.

Next step would be to add a test case for testing the read and write
operations.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>