Fix filters for is (not) empty/null
While refactoring generating the filter from a common method missed the
filterIfMissing setting for these conditions. They need to be hard coded
to function properly.

Change the return type of getCellFilter to return the exact type it creates
so that it can be easily modified by invokers as is the case here.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>