Add SPI for executing HBase operations in contained environment
The SPI basically abstracts the initialization of configuration, table pool
and creation of table. It also abstracts the exception wrapping from an
operation and returning a table to the pool. Thus allowing clients to
specifically concentrate on the main operation rather than working on its
surroundings. It uses callback handler to allow users to plugin their main
actions.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>