Extend filter config to be able to specify a cell range
It is particular required to be able to scan across a range of cells with
common prefix. An example would be authors of a book. A design could be
that authros is a column family and author:P are the qualifiers, where P is
author id and value of the qualifier is the author's name

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>