Make ID and Version of PersistentDTO template-able through generics
The main reason to do it is the use case of HBase where ID is a byte array
and version is a timestamp represented as a long. Similarly in other cases
it might be that ID and Version are different other data types. So we keep
it generic in the API as well, but with the constraint that they must be
Comparable.

As a result the old Persistent DTO is out of date but its being used in a
lot of places thus kept it intact but deprecated it to discourage future
use. In any case since it extends the new abstract class there is no extra
code to maintain.

Also had to modify the codes where stupidly extends and implements were
being used together.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>