Add support for merge when doing Put to a table
This is because HBase does a implicit merge of columns when doing a put.
So in order to ensure that unwanted cells do not exist, before Put-ting
those cells need to be deleted and thats what merge does. And it does only
on the family being put not all families. Merge implementations should be
wary of merging with non-existing rows which the diff merge implementation
takes care of.

In support for this 2 extra configurations are added to CommonDao a boolean
to control centrally whether merge should take place or not and the merge
service itself.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>