Allow dynamic column projection for HBase
In HBase column schema is not strict, a column family can hold any number
of columns with any qualifier. Now if we want to project a dynamic column
we need a way to inform HBase about it. So it would be 'familyConfigId:
qualifier', where ':' is the separator, specifying that it can not be used
in a name for family or qualifier.

The older mecahnism for projection is still available. It is possible to
project only a family or a static pre-deterministic column as well.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>