Make lock externally attainable
This change makes it possible to attain locks before even sending DTO to
DAO. This can be particularly useful to ensure lock is attained to perform
some test before ensuring to go ahead to DAO layer.

To be noted is though the default implementation uses weak hashmap clients
should invoke either one of evicts to ensure locks are not cached when next
time they are requested for the same instance.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>